Support for IPv6 server URL and transport
This patch enables support for connecting to an SSCEP server
using the IPv6 protocol. Specific changes include:

- Parse URL containing an IPv6 address.
- Replace call to gethostbyname() with getaddrinfo() for IPv6
  support in name resolution.

Prior to this patch, URL parsing was simply looking for a
colom (':') as port delimiter. This is incompatible with the
syntax for a numerical IPv6 address in a URL
('http://[fc00:1:2::3]:8080/'). The fix looks for a '['
character and if found, considers the following characters
to be an IPv6 address until reaching a ']', and then resumes
normal parsing.

In addition, gethostbyname() is replaced with getaddrinfo()
as the latter support IPv6 name resolution.