Get capabilities from SCEP server using GetCACaps
The GetCACaps operation provides a mechanism to retrieve the
SCEP capabilities of the CA. If supported the operation exposes a list
of keywords in plain text representing the supported capabilities. A
successful response is parsed and represented in a capabilities bitmask:
ca_caps.

See Section 3.5 - CA Capabilities for more information. Currently 9
such capabilities are defined in the specifications and recognised.

This is useful when the CA only supports POST requests for
'PKIOperation', which is represented by the POSTPKIOperation or
SCEPStandard capabilities. The POST method is more suitable for these
types of operation, as it avoids base64+URL-encoding data, caching and
possibly truncation of large HTTP request lines when using the GET
method otherwise.