Support POST method for PKI operations
According to the specifications [1], POST is the preferred method when
handling PKI operations while transferring binary data. Whether or not
the SCEP server supports the POST method is exposed by capabilities
operation.

If the POSTPKIOperation or SCEPStandard is advertised as supported by
the SCEP server, POST is used as the preferred method. This affects the
format of the message data of the PKI operation, as base64 encoding of
PKCS7 data is no longer necessary, nor is the subsequent URL-encoding.

[1] See section 2.9 - Mandatory-to-Implement Functionality.