Fix GetCRL/GetCert for situations where RA and CA certifificate differ
The GetCRL/GetCert commands send an issuer-serial to identify the
requested ressource. This information is read from the RA certificate
which will return unusable information when the SCEP RA certificate
is not the CA itself. This is a common situation when running SCEP
on a "real" CA and not in an embded setup.

This patch introduces the "-O" and "-s" flags to pass a different
certificate to read the required name and serial from.