Improve auto-select of certificates
If performing a GETCRL request and auto-selecting certificates, guess
at the issuer certificate by using the certificate that has the same
subject name as the issuer of the local certificate (specified with
'-l').
Previously, auto-selection would use the auto-selected 'cacert', which
would most likely _not_ be the issuing certificate of the local
certificate.

This may still be overridden by using the -O command line option to
specify the issuer certificate.