Close #62 #149: pass a private key passphrase to an enroll request
Implement the Openssl syntax according to:
https://www.openssl.org/docs/man3.0/man1/openssl-passphrase-options.html

Enable this feature (and keyboard login) for ENGINE keys.

Removing the passphrase from a software key is sometimes possible,
but removing it from from an ENGINE key is usually impossible.

The OS will drop the allocated memory for the UI_METHOD.
This is a short running app after all.

Signed-off-by: Christian Hohnstaedt <christian@hohnstaedt.de>